// src/shared/dataprovider/api/model/messageStatusEnum.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type MessageStatusEnum = 1 | 2 | 3;

export const MessageStatusEnum = {
    NUMBER_1: 1 as MessageStatusEnum,
    NUMBER_2: 2 as MessageStatusEnum,
    NUMBER_3: 3 as MessageStatusEnum
};