// src/shared/dataprovider/api/model/pushNotificationMessageDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { MessageStatusEnum } from './messageStatusEnum';

export interface PushNotificationMessageDTO { 
    Id?: number;
    UserId?: number;
    Message?: string;
    Title?: string;
    DateCreated?: Date;
    DateUpdated?: Date;
    Status?: MessageStatusEnum;
}